% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_charts.R
\name{create_all_charts}
\alias{create_all_charts}
\title{Function to create all charts within package}
\usage{
create_all_charts(data = load_data())
}
\arguments{
\item{data}{default is \code{load_data()}}
}
\value{
A list of ggplot2 objects
}
\description{
Function to create all charts within package
}
